Refactor color-related code and add color utility function
