+++ client/app/types/styles.ts
@@ -41,3 +41,24 @@ export enum buttonsTheme {
 export type buttonSize = 'small' | 'medium' | 'large'
 
 export type fontSize = 'xs' | 'sm' | 'md' | 'lg' | 'xl'
+
+/** Colors */
+
+export type HexColor = `#${string}`
+export type ColorGroup = 'RED' | 'GRAY' | 'CLOUD_BLUE'
+export type ColorShade = '100' | '200' | '300' | '400' | '500' | '600'
+export type BasicColor = 'WHITE' | 'BLACK'
+
+export type Colors = {
+  [K in `${ColorGroup}_${ColorShade}`]: HexColor
+} & {
+  [K in BasicColor]: HexColor
+}
+
+export type ColorsByGroup<T extends ColorGroup> = {
+  [K in `${T}_${ColorShade}`]: HexColor
+}
+
+export type BasicColors = {
+  [K in BasicColor]: HexColor
+}
