+++ client/app/utils/common.ts
@@ -1,5 +1,8 @@
 /** MouseEvent */
 
+import { ColorsByGroup, ColorGroup } from '@/app/types'
+import { COLORS } from '@/app/styles'
+
 interface DragHorizonProps {
   event: MouseEvent | React.MouseEvent<HTMLButtonElement, MouseEvent>
   leftCallback: () => void
@@ -42,4 +45,18 @@ export function changeToDate(time: Date): string {
 
 export function changeToTime(time: Date): string {
   return new Date(time.toString()).toLocaleString('ko')
+}
+
+/** Colors */
+
+export function getColorsByGroup<T extends ColorGroup>(color: T): ColorsByGroup<T> {
+  return Object.entries(COLORS)
+    .filter(([key]) => key.startsWith(`${color}_`))
+    .reduce(
+      (acc, [key, value]) => ({
+        ...acc,
+        [key]: value
+      }),
+      {} as ColorsByGroup<T>
+    )
 }
